Rename theme.js to defaultTheme.js
